feat: refactor: aaded icons
